+++ CHANGELOG.md
@@ -6,6 +6,8 @@ New:
  - Add an option to allow ICMP rules in default security groups
  - Alerts for max instances per user in project
  - Alerts for max vCPUs per user in project
+ - Alerts for max volumes per user in project
+ - Alerts for max storage per user in project
 
 Fixes:
  - Fix the instance migration issue in gerenuk-libvirtmon service